VisualStudio2010仮対応 Matrix43をMatrix34へ変更 行ベクトル計算から列ベクトル計算、ベクトルに左から行列を掛けるように統一

+++ trunk/lframework/scene/IOGeometry.cpp
@@ -50,7 +50,7 @@ namespace lscene
         GeomHeader header;
 
         header.id_ = GeometryID;
-        header.primCount_ = geom.getPrimitiveCount();
+        header.primCount_ = static_cast<u16>( geom.getPrimitiveCount() );
         header.indexOffset_ = geom.getIndexOffset();
         header.numIndices_ = geom.getNumIndices();
         header.materialIndex_ = geom.getMaterialIndex();
